[RQ-OSS-1050]: Move x-password from header to query params
